Refactor recipe handling in Create AI Recipe page
- Moved recipe data extraction logic to a new utility file for better organization.
- Enhanced the UI to include a "Saved" indicator for previously saved recipes.
- Updated the recipe addition handler to manage saved responses more effectively.
